requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6
Babel>=1.3
docker>=2.4.0
jinja2>=2.8
PyYAML>=3.11
future>=0.14.3
cliff>=1.10.1
